Got the app to show tweets of certain id!
The error was in type conversion - though the url specified an id, it
was a string, not a number. Hence when using the where function, the
lookup would fail because there was no id such as {id: (id as string)}.
Fixed by just converting id from url to # by multiplying by 1.